+++ src/main/kotlin/com/twilio/apkscale/tasks/MeasureAndroidLibrarySizeTask.kt
@@ -1,12 +1,11 @@
 package com.twilio.apkscale.tasks
 
+import com.android.build.api.variant.LibraryAndroidComponentsExtension
 import com.android.build.gradle.LibraryExtension
-import com.android.build.gradle.api.LibraryVariant
 import com.google.gson.Gson
 import com.twilio.apkscale.ApkscaleExtension
 import com.twilio.apkscale.model.ApkscaleReport
 import org.gradle.api.DefaultTask
-import org.gradle.api.DomainObjectSet
 import org.gradle.api.Project
 import org.gradle.api.artifacts.DependencySet
 import org.gradle.api.tasks.TaskAction
@@ -31,7 +30,21 @@ open class MeasureAndroidLibrarySizeTask @Inject constructor(
     companion object {
         const val MEASURE_TASK_NAME = "measureSize"
 
-        fun create(project: Project, libraryExtension: LibraryExtension, apkscaleExtension: ApkscaleExtension) {
+        fun create(
+            project: Project,
+            libraryExtension: LibraryExtension,
+            componentsExtension: LibraryAndroidComponentsExtension,
+            apkscaleExtension: ApkscaleExtension,
+        ) {
+            // register to capture variants
+            val variantDependencies = HashMap<String, DependencySet>()
+            componentsExtension.onVariants {
+                variantDependencies.put(
+                    it.name,
+                    it.compileConfiguration.allDependencies,
+                )
+            }
+            // after evaluation step assign
             project.afterEvaluate {
                 val measureTask = project.tasks.create(
                     MEASURE_TASK_NAME,
@@ -40,34 +53,23 @@ open class MeasureAndroidLibrarySizeTask @Inject constructor(
                     apkscaleExtension.humanReadable,
                     libraryExtension.defaultConfig.minSdkVersion?.apiLevel,
                     libraryExtension.defaultConfig.targetSdkVersion?.apiLevel,
-                    getVariantDependencies(libraryExtension.libraryVariants),
+                    variantDependencies,
                     libraryExtension.ndkVersion ?: "",
                 )
-
                 // Ensure that measure task runs after assemble tasks
                 measureTask.mustRunAfter(project.tasks.named("assemble"))
                 libraryExtension.buildTypes.forEach {
-                    measureTask.mustRunAfter(project.tasks.named("assemble${it.name.capitalize()}"))
+                    measureTask.mustRunAfter(project.tasks.named("assemble${it.name.capitalizeLocalAware()}"))
                 }
-                libraryExtension.libraryVariants.forEach {
-                    measureTask.mustRunAfter(project.tasks.named("assemble${it.name.capitalize()}"))
+                variantDependencies.forEach {
+                    measureTask.mustRunAfter(project.tasks.named("assemble${it.key.capitalizeLocalAware()}"))
                 }
-            }
-        }
-
-        private fun getVariantDependencies(libraryVariants: DomainObjectSet<LibraryVariant>): Map<String, DependencySet> {
-            /*
-             * Create a map of the library variants to the variant's dependencies so that apkscale pulls in the
-             * correct dependencies for each variant that is measured.
-             */
-            return libraryVariants.associate {
-                it.name.lowercase(Locale.getDefault()) to it.compileConfiguration.allDependencies
             }
         }
     }
 
-    private val outputAarDir = project.buildDir.resolve("outputs/aar")
-    private val apkscaleDir = File("${project.buildDir}/apkscale")
+    private val outputAarDir = project.layout.buildDirectory.dir("outputs/aar").get().asFile
+    private val apkscaleDir = project.layout.buildDirectory.dir("apkscale").get().asFile
     private val appMainDir = File("$apkscaleDir/src/main")
     private val apkscaleOutputDir = File("$apkscaleDir/build/outputs/reports")
     private val buildFile = File(apkscaleDir, "build.gradle")
@@ -197,7 +199,7 @@ open class MeasureAndroidLibrarySizeTask @Inject constructor(
                     mavenCentral()
                   }
                   dependencies {
-                    classpath 'com.android.tools.build:gradle:8.0.2'
+                    classpath 'com.android.tools.build:gradle:8.3.0'
                   }
                 }
                 apply plugin: 'com.android.application'
@@ -309,6 +311,13 @@ open class MeasureAndroidLibrarySizeTask @Inject constructor(
         return aarFileName.substringAfter("${project.name}-")
             .substringBefore(".aar")
             .replace("-", "")
-            .lowercase(Locale.getDefault())
+    }
+}
+
+fun String.capitalizeLocalAware(): String {
+    return this.replaceFirstChar {
+        if (it.isLowerCase()) {
+            it.titlecase(Locale.getDefault())
+        } else it.toString()
     }
 }